requirements.txt
# Discord Bot Dependencies
discord.py>=2.3.0
python-dotenv>=1.0.0
# SSH Client for Remote Server Management
paramiko>=4.0.0
